Move presetting methods to presetting.py
